use time for judge instead distance

+++ src/components/receptor/step_left.rs
@@ -32,9 +32,7 @@ impl PatternReceptor for StepLeftReceptor {
     }
 
     fn init_or_defer(&mut self, current_time: f64, bpm: f32) {
-        if (current_time - self.last_time).abs()
-            > bpm.recip() as f64 * 60.0 + ERROR_THRESHOLD as f64
-        {
+        if (current_time - self.last_time).abs() > bpm.recip() as f64 * 60.0 + ERROR_THRESHOLD {
             self.init();
         }
     }
